+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### Gazebo 6.X.X (201X-XX-XX)
 
+1. Backport model editor toolbar fixed joint option from [pull request #1794](https://bitbucket.org/osrf/gazebo/pull-request/1794)
+    * [Pull request #1957](https://bitbucket.org/osrf/gazebo/pull-request/1957)
+
 1. Fix visual transparency issues
     * [Pull request #1967](https://bitbucket.org/osrf/gazebo/pull-request/1967)
     * [Issue #1726](https://bitbucket.org/osrf/gazebo/issue/1726)